+++ doc/whats_new.rst
@@ -79,7 +79,7 @@ API
 
     - Allow passing a list of channel names as ``show_names`` in function  :func:`mne.viz.plot_sensors` and methods :meth:`mne.Evoked.plot_sensors`, :meth:`mne.Epochs.plot_sensors` and :meth:`mne.io.Raw.plot_sensors` to show only a subset of channel names by `Jaakko Leppakangas`_
 
-    - Make function `mne.io.eeglab.read_eeglab_events` public to allow loading overlapping events from EEGLAB files, by `Jona Sassenhagen`_.
+    - Make function `mne.io.eeglab.read_events_eeglab` public to allow loading overlapping events from EEGLAB files, by `Jona Sassenhagen`_.
 
 .. _changes_0_14: